+++ shapes_converter/src/shex_to_sparql_error.rs
@@ -1,5 +1,5 @@
 use prefixmap::IriRef;
-use shex_ast::{Schema, SchemaJsonError};
+use shex_ast::{Schema, SchemaJsonError, ShapeExprLabel};
 use thiserror::Error;
 
 #[derive(Error, Debug)]
@@ -7,9 +7,34 @@ pub enum ShEx2SparqlError {
     #[error("Shape {iri} not found in schema {schema:?}")]
     ShapeNotFound { iri: IriRef, schema: Schema },
 
+    #[error("Shape reference {sref} not found in schema {schema:?}")]
+    ShapeRefNotFound {
+        sref: ShapeExprLabel,
+        schema: Schema,
+    },
+
+    #[error("No shapes found in schema to convert to SPARQL. Schema\n{schema:?}")]
+    NoShapes { schema: Schema },
+
+    #[error(
+        "No shape found to convert to SPARQL because list of shapes is empty. Schema\n{schema:?}"
+    )]
+    EmptyShapes { schema: Schema },
+
     #[error(transparent)]
     SchemaError {
         #[from]
         err: SchemaJsonError,
     },
+
+    #[error("ShEx2Sparql: Feature not implemented: {msg}")]
+    NotImplemented { msg: String },
+}
+
+impl ShEx2SparqlError {
+    pub fn not_implemented(msg: &str) -> ShEx2SparqlError {
+        ShEx2SparqlError::NotImplemented {
+            msg: msg.to_string(),
+        }
+    }
 }